final signup changes added and title switched to image
